+++ swabbie-core/src/main/kotlin/com/netflix/spinnaker/swabbie/tagging/AwsTag.kt
@@ -31,3 +31,20 @@ data class UpsertImageTagsRequest(
   override val application: String,
   override val description: String
 ) : ImageTagsRequest("upsertImageTags", application, description)
+
+open class AmazonServerGroupTagsRequest(
+  open val type: String,
+  open val application: String,
+  open val description: String
+) : TagRequest
+
+data class UpsertServerGroupTagsRequest(
+  val serverGroupName: String,
+  val regions: Set<String>,
+  val tags: Map<String, String>,
+  val cloudProvider: String,
+  val cloudProviderType: String,
+  val credentials: String,
+  override val application: String,
+  override val description: String
+) : AmazonServerGroupTagsRequest("upsertServerGroupTags", application, description)
